+++ packages/app/src/routes/(app)/groups/[uid]/+page.svelte
@@ -102,11 +102,14 @@
   $: ({ group } = data);
   $: ShopItems = data.itemsOfGroup;
 
-  $: canEditArticles = Boolean($me?.admin || myPermissions?.canEditArticles || meOnClubBoard);
+  $: canEditArticles = Boolean(
+    $me?.admin || myPermissions?.canEditArticles || meOnClubBoard || data.canEditGroup,
+  );
   $: canEditEvents = canEditArticles;
   $: canEditMembers = Boolean(
     $me?.admin || myPermissions?.canEditMembers || meOnClubBoard || data.canEditGroup,
   );
+  $: canEditDetails = Boolean($me?.admin || group?.canEditDetails || data.canEditGroup);
 
   const joinGroup = async (groupUid: string) => {
     if (!$me) return goto(`/login?${new URLSearchParams({ to: $page.url.pathname }).toString()}`);
@@ -180,7 +183,7 @@
       <h1>
         {group.name}
         <ButtonShare />
-        {#if group.canEditDetails}
+        {#if canEditDetails}
           <ButtonGhost help="Modifier les infos" href="./edit/"><IconGear /></ButtonGhost>
         {/if}